+++ src/modules/ProgrammeManager/AppShell.jsx
@@ -4,6 +4,7 @@ import { ViewProvider } from './context/ViewContext';
 import { CalendarProvider } from './context/CalendarContext';
 import { SelectionProvider } from './context/SelectionContext';
 import { FilterProvider } from './context/FilterContext';
+import { LayoutProvider } from './context/LayoutContext';
 import RibbonTabs from './components/RibbonTabs/RibbonTabs';
 import SidebarTree from './components/SidebarTree';
 import TaskGrid from './components/TaskGrid';
@@ -12,6 +13,7 @@ import TaskPropertiesPane from './components/TaskPropertiesPane';
 
 function AppShellContent() {
   const sidebarRef = useRef();
+  const { getPaneSize } = useLayoutContext();
 
   const handleExpandAll = () => {
     sidebarRef.current?.expandAll?.();
@@ -33,24 +35,38 @@ function AppShellContent() {
 
         {/* Main Content Area - Flex row with three panes */}
         <div className='flex-1 flex overflow-hidden'>
-          {/* SidebarTree - Left pane (~20%) */}
-          <div className='w-1/5 min-w-[200px] max-w-[300px] bg-white border-r border-gray-300 overflow-auto'>
+          {/* SidebarTree - Left pane */}
+          <div 
+            className='bg-white border-r border-gray-300 overflow-auto'
+            style={{ 
+              width: getPaneSize('sidebar') === 'flex-1' ? '300px' : getPaneSize('sidebar'),
+              minWidth: '200px',
+              maxWidth: '400px'
+            }}
+          >
             <SidebarTree ref={sidebarRef} />
           </div>
 
-          {/* TaskGrid - Middle pane (~40%) */}
-          <div className='w-2/5 min-w-[300px] bg-white border-r border-gray-300 overflow-auto'>
+          {/* TaskGrid - Middle pane */}
+          <div className='flex-1 bg-white border-r border-gray-300 overflow-auto'>
             <TaskGrid />
           </div>
 
-          {/* GanttChart - Right pane (~40%) */}
+          {/* GanttChart - Right pane */}
           <div className='flex-1 bg-white overflow-auto'>
             <GanttChart />
           </div>
         </div>
 
         {/* TaskPropertiesPane - Bottom pane, full width */}
-        <div className='h-64 min-h-[200px] max-h-[400px] bg-white border-t border-gray-300 overflow-auto'>
+        <div 
+          className='bg-white border-t border-gray-300 overflow-auto'
+          style={{ 
+            height: getPaneSize('properties') === 'flex-1' ? '256px' : getPaneSize('properties'),
+            minHeight: '200px',
+            maxHeight: '400px'
+          }}
+        >
           <TaskPropertiesPane />
         </div>
       </div>
@@ -60,14 +76,16 @@ function AppShellContent() {
 
 export default function AppShell() {
   return (
-    <FilterProvider>
-      <SelectionProvider>
-        <CalendarProvider>
-          <ViewProvider>
-            <AppShellContent />
-          </ViewProvider>
-        </CalendarProvider>
-      </SelectionProvider>
-    </FilterProvider>
+    <LayoutProvider>
+      <FilterProvider>
+        <SelectionProvider>
+          <CalendarProvider>
+            <ViewProvider>
+              <AppShellContent />
+            </ViewProvider>
+          </CalendarProvider>
+        </SelectionProvider>
+      </FilterProvider>
+    </LayoutProvider>
   );
 }